requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
constantly==23.10.4
cryptography==42.0.7
Django==4.2.7
django-cors-headers==4.3.1
django-modeltranslation==0.18.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2023.3.post1
PyYAML==6.0.1
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0
zope.interface==6.4.post2
